reorganize mango query class with a selector module
